fixed the rotation glitch
